requirements.txt
face_recognition==1.2.2
Pillow==5.1.0
numpy==1.14.1
nmslib==1.7.2
google-api-python-client==1.6.7
